+++ aclients/jsonrpc/aio_jrpc_server.py
@@ -6,7 +6,6 @@
 @software: PyCharm
 @time: 2020/2/21 下午5:00
 """
-from typing import Optional
 
 from sanic import Sanic
 
@@ -18,17 +17,38 @@ except ImportError as e:
 __all__ = ("SanicJsonRPC",)
 
 
-class SanicJsonRPC(Jsonrpc):
+class SanicJsonRPC(object):
+    """
+    sanic jsonrpc object
     """
 
-    """
-
-    def __init__(self, app: Sanic):
+    def __init__(self, app: Sanic = None, post_route: str = "/api/jrpc/post", ws_route: str = "/api/jrpc/ws"):
         """
-
+        jsonrpc 实例初始化
         Args:
+            app: app应用
+            post_route: post url
+            ws_route: websocket url
 
         """
-        post_route: Optional[str] = "/api/jrpc/post"
-        ws_route: Optional[str] = "/api/jrpc/ws"
-        super().__init__(app, post_route, ws_route)
+        self.jrpc: Jsonrpc = None
+        self.post_route: str = post_route
+        self.ws_route: str = ws_route
+
+        if app is not None:
+            self.init_app(app)
+
+    def init_app(self, app: Sanic = None, post_route: str = "/api/jrpc/post", ws_route: str = "/api/jrpc/ws"):
+        """
+        jsonrpc 实例初始化
+        Args:
+            app: app应用
+            post_route: post url
+            ws_route: websocket url
+        Returns:
+
+        """
+        self.post_route = post_route or self.post_route
+        self.ws_route = ws_route or self.ws_route
+
+        self.jrpc = Jsonrpc(app, post_route=self.post_route, ws_route=self.ws_route)